// userAndStrategyService/enums/messages.ts
export enum SuccessMessages {
    BROKER_INTEGRATION = 'Integration is complete',
    OPERATION_SUCCESSFUL = 'Operation completed successfully.',
    RESOURCE_CREATED = 'Resource created successfully.',
    USER_SIGNED_UP = 'User account created successfully.',
    LOGIN_SUCCESSFUL = 'Login successful.',
    USER_UPDATED = 'User profile updated successfully.',
    USER_DELETED = 'User account deleted successfully.',
    LOGOUT_SUCCESSFUL = 'Logged out successfully.',
    ORDER_PLACED = 'Order successfully placed.',
    ORDER_MODIFIED = 'Order successfully modified.',
    ORDER_CANCELED = 'Order successfully canceled.',
    STRATEGY_STARTED = 'Trading strategy started successfully.',
    ACCOUNT_FUNDED = 'Account funded successfully.',
    HEALTH_CHECK_OK = 'Service is operational.',
    UPSTOX_AUTH_URL_GENERATED = 'Upstox authentication URL generated successfully.',
    UPSTOX_ACCESS_TOKEN_GENERATED = 'Upstox access token generated and stored successfully.',
    UPSTOX_WEBSOCKET_ESTABLISHED = 'Upstox market data websocket connection established.',
    UPSTOX_INSTRUMENT_SUBSCRIBED = 'Successfully subscribed to the instrument on Upstox.',
    UPSTOX_INSTRUMENT_DETAILS_FETCHED = 'Instrument details fetched successfully from Upstox.',
    INSTRUMENT_UNSUBSCRIBED = 'Instrument unsubscribed successfully',
    BROKER_INTEGRATION_UPDATED = 'Integration update successfully',
    BROKER_INTEGRATION_DELETED = 'integration deleted successfully',
}